move serialize state function from env sampler to state template
